refactor: Move code to appropriate folders for CODEOWNERS
