+++ cloud_pipelines_backend/search/filter_query_sql.py
@@ -6,8 +6,8 @@ from typing import Any, Final
 
 import sqlalchemy as sql
 
-from . import backend_types_sql as bts
-from . import errors
+from .. import backend_types_sql as bts
+from .. import errors
 from . import filter_query_models
 
 SYSTEM_KEY_PREFIX: Final[str] = "system/"
